requirements.txt
skimage.metrics
imutils
cv2
PIL
numpy
streamlit